Convert modifier to function to minimize contract
