+++ contracts/accessors/SimulateTxAccessor.sol
@@ -1,4 +1,4 @@
-pragma solidity >=0.7.0 <0.8.0;
+pragma solidity >=0.7.0 <0.9.0;
 
 import "../base/Executor.sol";
