+++ contracts/handler/HandlerContext.sol
@@ -1,5 +1,5 @@
 // SPDX-License-Identifier: LGPL-3.0-only
-pragma solidity >=0.7.0 <0.8.0;
+pragma solidity >=0.7.0 <0.9.0;
 
 /// @title Handler Context - allows to extract calling context
 /// @author Richard Meissner - <richard@gnosis.pm>
