+++ contracts/interfaces/ERC1155TokenReceiver.sol
@@ -1,5 +1,5 @@
 // SPDX-License-Identifier: LGPL-3.0-only
-pragma solidity >=0.7.0 <0.8.0;
+pragma solidity >=0.7.0 <0.9.0;
 
 /**
     Note: The ERC-165 identifier for this interface is 0x4e2312e0.
