+++ contracts/test/ERC1155Token.sol
@@ -1,5 +1,5 @@
 // SPDX-License-Identifier: LGPL-3.0-only
-pragma solidity >=0.7.0 <0.8.0;
+pragma solidity >=0.7.0 <0.9.0;
 
 import "../interfaces/ERC1155TokenReceiver.sol";
 import "../external/GnosisSafeMath.sol";
